+++ app/gui/src/dashboard/hooks/backendBatchedHooks.ts
@@ -151,6 +151,35 @@ export function copyAssetsMutationOptions(backend: Backend) {
   })
 }
 
+/** The type of a "move assets" mutation. */
+type CopyAssetsMutation = Mutation<
+  null,
+  Error,
+  readonly [ids: readonly AssetId[], parentId: DirectoryId]
+>
+
+/** Return matching in-flight "move assets" mutations. */
+export function useCopyAssetsMutationState<Result>(
+  backend: Backend,
+  options: {
+    predicate?: (mutation: CopyAssetsMutation) => boolean
+    select?: (mutation: CopyAssetsMutation) => Result
+  } = {},
+) {
+  const { predicate, select } = options
+  return useMutationState({
+    filters: {
+      ...copyAssetsMutationOptions(backend),
+      predicate: (mutation: CopyAssetsMutation) =>
+        mutation.state.status === 'pending' && (predicate?.(mutation) ?? true),
+    },
+    // This is UNSAFE when the `Result` parameter is explicitly specified in the
+    // generic parameter list.
+    // eslint-disable-next-line no-restricted-syntax
+    select: select as (mutation: Mutation<unknown, Error, unknown, unknown>) => Result,
+  })
+}
+
 /** Call "move" mutations for a list of assets. */
 export function moveAssetsMutationOptions(backend: Backend) {
   return mutationOptions({
@@ -180,6 +209,35 @@ export function moveAssetsMutationOptions(backend: Backend) {
       ],
       awaitInvalidates: true,
     },
+  })
+}
+
+/** The type of a "move assets" mutation. */
+type MoveAssetsMutation = Mutation<
+  null,
+  Error,
+  readonly [ids: readonly AssetId[], parentId: DirectoryId]
+>
+
+/** Return matching in-flight "move assets" mutations. */
+export function useMoveAssetsMutationState<Result>(
+  backend: Backend,
+  options: {
+    predicate?: (mutation: MoveAssetsMutation) => boolean
+    select?: (mutation: MoveAssetsMutation) => Result
+  } = {},
+) {
+  const { predicate, select } = options
+  return useMutationState({
+    filters: {
+      ...moveAssetsMutationOptions(backend),
+      predicate: (mutation: MoveAssetsMutation) =>
+        mutation.state.status === 'pending' && (predicate?.(mutation) ?? true),
+    },
+    // This is UNSAFE when the `Result` parameter is explicitly specified in the
+    // generic parameter list.
+    // eslint-disable-next-line no-restricted-syntax
+    select: select as (mutation: Mutation<unknown, Error, unknown, unknown>) => Result,
   })
 }