+++ engine/runtime-integration-tests/src/test/scala/org/enso/interpreter/test/instrument/InstrumentTestContext.scala
@@ -40,6 +40,9 @@ abstract class InstrumentTestContext(packageName: String) {
   protected val runtimeServerEmulator: RuntimeServerEmulator =
     new RuntimeServerEmulator(messageQueue, lockManager)
 
+  final def send(msg: Api.Request): Unit =
+    runtimeServerEmulator.sendToRuntime(msg)
+
   def receiveNone: Option[Api.Response] = {
     Option(messageQueue.poll())
   }
@@ -112,7 +115,7 @@ abstract class InstrumentTestContext(packageName: String) {
   ): List[Api.Response] = {
     var count: Int                     = n
     var lastSeen: Option[Api.Response] = None
-    Iterator
+    val collected = Iterator
       .continually(receiveWithTimeout(timeoutSeconds))
       .filter(f)
       .takeWhile {
@@ -127,7 +130,13 @@ abstract class InstrumentTestContext(packageName: String) {
       }
       .flatten
       .filter(excludeLibraryLoadingPayload)
-      .toList ++ lastSeen
+      .toList
+
+    if (lastSeen.isEmpty || lastSeen == collected.lastOption) {
+      collected
+    } else {
+      collected ++ lastSeen
+    }
   }
 
   private def excludeLibraryLoadingPayload(response: Api.Response): Boolean =
@@ -137,6 +146,12 @@ abstract class InstrumentTestContext(packageName: String) {
       case _ =>
         true
     }
+
+  final def writeMain(contents: String): File =
+    Files.write(pkg.mainFile.toPath, contents.getBytes).toFile
+
+  final def executionComplete(contextId: java.util.UUID): Api.Response =
+    Api.Response(Api.ExecutionComplete(contextId))
 
   def close(): Unit = {
     if (context != null) {